requirements.txt
asgiref==3.9.1
Django==5.2.6
django-extensions==4.1
django-request-logging==0.7.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
